requirements.txt
kafka-python-ng
requests
opensearch-py